the data type of pwm_time should be "float".
Seem to be correct saved data in the SPI NOR FLASH

+++ SI1120.h
@@ -6,7 +6,7 @@ sbit 			SI1120_MD	=	P1^4;
 sbit 			SI1120_STX =	P1^5;	
 sbit 			SI1120_PRX	=	P3^2;	//pin INT0, trigger on both rising and falling edge
 // declaration pwm time as an external variable
-extern unsigned int pwm_time,pwm_time_min,pwm_time_max;
+extern float pwm_time,pwm_time_min,pwm_time_max;
 
 void SI1120_CONFIG_MODE_VIRH(void);
 void Enable_EXTINT0(void);
